+++ modules/blackhole/cuda/blackhole_cuda.h
@@ -2,6 +2,6 @@
 #define __OPENSPACE_MODULE_BLACKHOLE___CUDA___H__
 #include <vector>
 
-void schwarzchild(double rs, double envmap_r, int num_paths, int num_steps, double u_0, double h, double* angle_out);
+void schwarzchild(float rs, float envmap_r, int num_paths, int num_steps, float u_0, float h, float* angle_out);
 
 #endif
